% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbnormalMark.R, R/labFun.R
\name{getAbnormalMark}
\alias{getAbnormalMark}
\alias{abMark}
\title{Get abnormal mark}
\usage{
getAbnormalMark(
  labData,
  idColName,
  labItemColName,
  valueColName,
  genderColName,
  genderTable,
  referenceTable = refLOINC
)
}
\arguments{
\item{labData}{a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.}

\item{idColName}{the column name that records patient ID in labData.}

\item{labItemColName}{the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add \code{+} operator between column names, e.g., \code{A + B}.}

\item{valueColName}{the column name that records test value in labData. Only numeric value is acceptable.}

\item{genderColName}{the column name that records gender of each patient. Gender should be coded as \code{M} or \code{F}.}

\item{referenceTable}{a data table that records reference values of each lab item. Column names should be: \code{LowerBound_Male}, \code{UpperBound_Male}, \code{LowerBound_Female}, \code{UpperBound_Female}. The \code{labItemColName} of \code{labData} and \code{refLOINC} have to be the same.}
}
\value{
A \code{data.table} merged two tables.
}
\description{
\code{getAbnormalMark} generates a new column recorded abnormal values: high \code{H} and \code{L}.
}
\details{
This step is optional but recommended. Reference ranges differ in different institutions, so a user-defined reference table is needed. The table should record the reference values for male and female separately. Lower bound can be expressed as \code{<}, \code{(} (greater than) or  \code{<=} , \code{[} (great tham or euqal to), and the expression of upper bound is similar. If there is no upper bound or lower bound, just remain empty. After the new abnormal flag column generated, it can be used to do categorical analysis, or be used in the function: \code{plotTimeSeriesLab}.
}
\examples{
loincSample <- mapLOINC(labData = labSample,
                       labItemColName = ITEMID,
                       mappingTable = mapSample)

#Use LONIC table (default)
getAbnormalMark(labData = loincSample,
               idColName = SUBJECT_ID,
               labItemColName = LOINC,
               valueColName = VALUENUM,
               genderColName = GENDER,
               genderTable = patientSample,
               referenceTable = refLOINC)

#Use user define table
getAbnormalMark(labData = labSample,
               idColName = SUBJECT_ID,
               labItemColName = ITEMID,
               valueColName = VALUENUM,
               genderColName = GENDER,
               genderTable = patientSample,
               referenceTable = refUserDefine)

}
