% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Box Plot}
\usage{
box_plot(df, column)
}
\arguments{
\item{df}{a dataframe}

\item{column}{a string for a variable in the dataframe df}
}
\value{
A box plot of \code{column} versus DRG codes.
}
\description{
Box Plot
}
\examples{
DRG <- read_csv('DRG_data.csv')

box_plot(DRG, 'Average Medicare Payments')

}
