% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots - DRG.R
\name{boxplot_DRG}
\alias{boxplot_DRG}
\title{a function to make a boxplot of payments by DRG code.}
\usage{
boxplot_DRG(data, payment)
}
\arguments{
\item{data}{a dataframe that we want to analysis (DRG)}

\item{payment}{Choose average medicare payments,the average total payment, or the average covered charges as the y axis for the plot}
}
\value{
output a boxplot with the average payment you choose
}
\description{
a function to make a boxplot of payments by DRG code.
}
\examples{
boxplot_DRG(DRG,'Average.Medicare.Payments')
}
