% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot_function.R
\name{boxplot_function}
\alias{boxplot_function}
\title{Boxplot Function}
\usage{
boxplot_function(df, vary)
}
\arguments{
\item{df}{a dataframe}

\item{vary}{a string name for variable y in the dataframe df}
}
\value{
A plot with boxplots of each DRG code versus \code{vary}
}
\description{
This function produces a ggplot boxplot of DRG code versus \code{vary}
}
\examples{
DRG code versus average Medicare payments
boxplot_function(data, Average.Medicare.Payments)


}
