% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2function.R
\name{boxplotter}
\alias{boxplotter}
\title{Boxplot of payments by DRG code}
\usage{
boxplotter(payment_type)
}
\arguments{
\item{payment_type}{a string for type of payments}
}
\value{
A boxplot of payments by DRG code
}
\description{
This function makes a boxplot of payments by DRG code.
}
\examples{
boxplotter('Average Covered Charges')

}
