% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack problem using brute force knapsack algortihm}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data.frame with two variables v and w (values and weights)}

\item{W}{size (Max weight capacity)}

\item{parallel}{bool}
}
\value{
the maximum values and elements to be carried by knapsack
}
\description{
Knapsack problem using brute force knapsack algortihm
}
