% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack_general.R
\name{brute_force_knapsack_general}
\alias{brute_force_knapsack_general}
\title{generlized Brute Force Knapsack Solver}
\usage{
brute_force_knapsack_general(x, W, parallel)
}
\arguments{
\item{x}{A data frame containing two columns: \code{v} (values) and \code{w} (weights) for each item.}

\item{W}{A numeric value representing the maximum capacity (weight) of the knapsack.}

\item{parallel}{A logical value indicating whether to enable parallel processing.}
}
\value{
A list with the following components:
\item{value}{The maximum value that can be obtained.}
\item{elements}{The indices of the items selected to obtain the maximum value.}
\item{weight}{The total weight of the selected items.}
\item{parallel}{True if user wants to run brute force with parallel, false if user wants regular brute force}
}
\description{
Solves the knapsack problem using a parallelized brute force approach or with non paralellalized brute force approach depending on inpur.
}
\details{
This function evaluates calls brute_force_knapsack or brute_force_knapsack_optimized depending on boolean input for parameter parallel
}
