% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labFun.R
\name{commonLabArgs}
\alias{commonLabArgs}
\title{Common arguments of laboratory processing functions}
\arguments{
\item{labData}{a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.}

\item{idColName}{the column name that records patient ID in labData.}

\item{labItemColName}{the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add \code{+} operator between column names, e.g., \code{A + B}.}

\item{dateColName}{the column name that records test date in labData. It should be in \code{"YYYYMMDD"/"YYYY-MM-DD"} format.}

\item{indexDate}{the specific date that used for cutting time window. It can be first record (\code{"First"}), last record (\code{"Last"}), or any date of interest with \code{"YYYYMMDD"/"YYYY-MM-DD"} format.}

\item{gapDate}{desired period (in days) of each window interval. If \code{NULL}, it will be seen as only one single time window.}

\item{completeWindows}{logical. If \code{TRUE}, time window series will be complete in order. If \code{TRUE}, a window without lab test will be skipped. Default is \code{TRUE}.}
}
\description{
Common arguments of laboratory processing functions
}
