% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{count_stars_P}
\alias{count_stars_P}
\title{count_stars_P
It rates P-value with stars}
\usage{
count_stars_P(pval)
}
\arguments{
\item{pval}{The P_value}
}
\value{
0 "\emph{\strong{" 0.001 "}" 0.01 "}" 0.05 "." 0.1 " " 1
}
\description{
count_stars_P
It rates P-value with stars
}
