% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(graph, init_value)
}
\arguments{
\item{graph}{A data frame}

\item{init_value}{A number}
}
\value{
A vector with distances
}
\description{
Shortest path between nodes in a graph
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstras_algorithm}
}
