% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_stats_function.R
\name{DRG_stats_function}
\alias{DRG_stats_function}
\title{Statistics of DRG Codes Function}
\usage{
DRG_stats_function(df, statistic)
}
\arguments{
\item{df}{a dataframe}

\item{statistic}{a test statistic (ie, mean, median)}
}
\value{
A table with DRG code and \code{statistic} of average Medicare payments
}
\description{
This function produces a table of DRG codes and the corresponding test statistic of interest (ie, mean, median) of average Medicare payments
}
\examples{
DRG codes by mean average Medicare payments
DRG_stats_function(data, mean)


}
