% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The GCD of \code{a} and \code{b}
}
\description{
Algorithm to look for the GCD of two numbers
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
