\name{FahadUgurcanLab4-package}
\alias{FahadUgurcanLab4-package}
\alias{FahadUgurcanLab4}
\docType{package}
\title{
\packageTitle{FahadUgurcanLab4}
}
\description{
\packageDescription{FahadUgurcanLab4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FahadUgurcanLab4}
\packageIndices{FahadUgurcanLab4}

}
\author{
\packageAuthor{FahadUgurcanLab4}

Maintainer: \packageMaintainer{FahadUgurcanLab4}
}
\references{

}

\keyword{ package }
\seealso{


}
\examples{

}
