% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_crime_from_time.R
\name{find_crime_from_month}
\alias{find_crime_from_month}
\title{find crime from time}
\usage{
find_crime_from_month(data, x)
}
\arguments{
\item{data}{get crime data}

\item{x}{month input}
}
\description{
find crime from time
}
