% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2find_from_time_and_type.R
\name{find_crime_from_time_and_type}
\alias{find_crime_from_time_and_type}
\title{find crime from both time and type}
\usage{
find_crime_from_time_and_type(data, type, x)
}
\arguments{
\item{data}{get crime data}

\item{type}{crime type input}

\item{x}{month input}
}
\description{
find crime from both time and type
}
