% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeriesLab.R, R/labFun.R
\name{getTimeSeriesLab}
\alias{getTimeSeriesLab}
\alias{getTS}
\title{Obtaining Time-Series Data}
\usage{
getTimeSeriesLab(
  labData,
  idColName,
  labItemColName,
  dateColName,
  valueColName,
  indexDate = last,
  gapDate = NULL,
  completeWindows = TRUE
)
}
\arguments{
\item{labData}{a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.}

\item{idColName}{the column name that records patient ID in labData.}

\item{labItemColName}{the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add \code{+} operator between column names, e.g., \code{A + B}.}

\item{dateColName}{the column name that records test date in labData. It should be in \code{"YYYYMMDD"/"YYYY-MM-DD"} format.}

\item{valueColName}{the column name that records test value in labData. Only numeric value is acceptable.}

\item{indexDate}{the specific date that used for cutting time window. It can be first record (\code{"First"}), last record (\code{"Last"}), or any date of interest with \code{"YYYYMMDD"/"YYYY-MM-DD"} format.}

\item{gapDate}{desired period (in days) of each window interval. If \code{NULL}, it will be seen as only one single time window.}

\item{completeWindows}{logical. If \code{TRUE}, time window series will be complete in order. If \code{TRUE}, a window without lab test will be skipped. Default is \code{TRUE}.}
}
\value{
A \code{data.table} with statistical summary.
}
\description{
\code{getTimeSeriesLab} seperates original data into multiple time windows, and summarize statistical information.
}
\details{
This function is used for seperating lab data into multiple time windows, and it provides overall statistical information: total count, maximun value, minimun value, mean, nearest record to index date of each time window. If \code{indexDate} is first, then it will be the earliest test date among all the lab tests.
}
\examples{
getTimeSeriesLab(labData = labMIMIC,
                idColName = SUBJECT_ID,
                labItemColName = ITEMID,
                dateColName = CHARTTIME,
                valueColName = VALUENUM,
                indexDate = first,
                gapDate = 360,
                completeWindows = TRUE)
}
