% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goose_race.R
\name{goose_race}
\alias{goose_race}
\title{Build an experiment of a goose moving around a grid.
This particular function will create the grid with certain elements blocking the path.}
\usage{
goose_race(
  x = 10,
  y = 10,
  n = 30,
  p = c(0.1, 0.99),
  starting_point = c(1, 1),
  ending_point = c(100, 100)
)
}
\arguments{
\item{x}{integer width of the grid}

\item{y}{integer height of the grid}

\item{n}{integer number of elements blocking path}

\item{p}{vector of probability of finding an obstacle and not}

\item{starting_point}{vector of length two, e.g., \code{c(x, y)} for where to start the goose}

\item{ending_point}{vector of length two, e.g., \code{c(x, y)} for where to end the race}
}
\value{
A tibble containing the state of the goose race
}
\description{
Build an experiment of a goose moving around a grid.
This particular function will create the grid with certain elements blocking the path.
}
