% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack problem using greedy algortihm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v and w (values and weights)}

\item{W}{size (Max weight capacity)}
}
\value{
the maximum values and elements to be carried by knapsack
}
\description{
Knapsack problem using greedy algortihm
}
