% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack_optimized.R
\name{greedy_knapsack_optimized}
\alias{greedy_knapsack_optimized}
\title{Uses a greedy algorithm to solve the knapsack problem by selecting items based on the highest value-to-weight ratio.
This is an approximation and may not always yield the optimal solution.}
\usage{
greedy_knapsack_optimized(x, W)
}
\arguments{
\item{x}{A data frame containing two columns: v (values) and w (weights).}

\item{W}{A numeric value representing the maximum weight capacity of the knapsack.}
}
\value{
A list with the following components:
\item{value}{The total value of the selected items.}
\item{elements}{The indices of the selected items that maximize the value.}
}
\description{
Uses a greedy algorithm to solve the knapsack problem by selecting items based on the highest value-to-weight ratio.
This is an approximation and may not always yield the optimal solution.
}
