% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeKNN.R, R/labFun.R
\name{imputeKNN}
\alias{imputeKNN}
\alias{impKNN}
\title{KNN Imputation}
\usage{
imputeKNN(labData, idColName, k)
}
\arguments{
\item{labData}{a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.}

\item{idColName}{the column name that records patient ID in labData.}

\item{k}{the numbers of selected nearestZneighbors.}
}
\value{
A new, wide-formatted imputed \code{data.table}.
}
\description{
\code{imputeKNN} uses k-nearest neighbors (kNN) method to fill missing values.
}
\details{
Given wide format data, data can conduct KNN imputation. If lab item code are composed by multiple columns, it will automatically combined as one. All other columns will all be seen as value columns.
}
\examples{

timeSeriesData <- getTimeSeriesLab(labData = labSample,
                                   idColName = SUBJECT_ID,
                                   labItemColName = ITEMID,
                                   dateColName = CHARTTIME,
                                   valueColName = VALUENUM,
                                   indexDate = first,
                                   gapDate = 360,
                                   completeWindows = TRUE)
wideData <- wideTimeSeriesLab(labData = timeSeriesData,
                  idColName = ID,
                  labItemColName = ITEMID,
                  windowColName = Window,
                  valueColName = Max)

imputeKNN(labData = wideData,
          idColName = ID + Window,
          k = 2)

}
