% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack problem using greedy algortihm}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v and w (values and weights)}

\item{W}{size (Max weight capacity)}
}
\value{
the maximum values and elements to be carried by knapsack
}
\description{
Knapsack problem using greedy algortihm
}
