% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic_programming.R
\name{knapsack_dynamic_programming}
\alias{knapsack_dynamic_programming}
\title{Dynamic Programming Knapsack Solver}
\usage{
knapsack_dynamic_programming(x, W)
}
\arguments{
\item{x}{A data frame containing two columns: \code{v} (values) and \code{w} (weights).}

\item{W}{A numeric value representing the maximum capacity (weight) of the knapsack.}
}
\value{
A list with the following components:
\item{max_value}{The maximum value that can be obtained.}
\item{selected_items}{The indices of the items selected to obtain the maximum value.}
}
\description{
Solves the knapsack problem using dynamic programming to find the optimal
selection of items that maximize the value within the given weight capacity.
}
