\name{knapsackpack-package}
\alias{knapsackpack-package}
\alias{knapsackpack}
\docType{package}
\title{
\packageTitle{knapsackpack}
}
\description{
\packageDescription{knapsackpack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsackpack}
\packageIndices{knapsackpack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{knapsackpack}

Maintainer: \packageMaintainer{knapsackpack}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
