% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labData.R
\docType{package}
\name{lab-package}
\alias{lab}
\alias{lab-package}
\title{lab: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\author{
\strong{Maintainer}: Chen-Ju Chen \email{ju.chin.0902@gmail.com}

Authors:
\itemize{
  \item Yi-Ju Tseng \email{yjtseng@mail.cgu.edu.tw}
}

}
\keyword{internal}
