% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3-package.R
\docType{package}
\name{lab3}
\alias{lab3}
\alias{lab3-package}
\alias{lab3-package}
\title{lab3: A package for computating the notorious dijkstra method.}
\arguments{
\item{graph}{A data frame}

\item{init_value}{A number}

\item{graph}{A data frame}

\item{init_value}{A number}

\item{a}{A number}

\item{b}{A number}
}
\value{
A vector with distances
#'Dijkstra's Algorithm 

Shortest path between nodes in a graph

A vector with distances

The GCD of \code{a} and \code{b}
}
\description{
The lab3 package provides two categories of important functions:
euclidean and dijkstra.
}
\section{lab3 functions}{

The lab3 functions ...
}

\references{
\url{https://en.wikipedia.org/wiki/Dijkstras_algorithm}
#'Euclidean Algorithm 

Algorithm to look for the GCD of two numbers

\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
