\name{Lab3Algorithms-package}
\alias{Lab3Algorithms-package}
\alias{Lab3Algorithms}
\docType{package}
\title{
Lab 3 Algorithms
}
\description{
This package contains the algorithms euclidean() and dijkstra() to solve Lab 3.
}
\author{
Juan Manuel Pardo and Nikolaus Stratil-Sauer
}
\references{
Euclidean algorithm: \url{https://en.wikipedia.org/wiki/Euclidean_algorithm} 
Dijkstra algorithm: \url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\examples{
euclidean(100,1000) returns the gcd of 10
}
