\name{Lab4pack-package}
\alias{Lab4pack-package}
\alias{Lab4pack}
\docType{package}
\title{
\packageTitle{Lab4pack}
}
\description{
\packageDescription{Lab4pack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab4pack}
\packageIndices{Lab4pack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab4pack}

Maintainer: \packageMaintainer{Lab4pack}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
