\name{lab6-package}
\alias{lab6-package}
\alias{lab6}
\docType{package}
\title{
\packageTitle{lab6}
}
\description{
\packageDescription{lab6}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{lab6}

Maintainer: \packageMaintainer{lab6}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = lab6)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{lab6}}
%\section{Documentation Index}{\packageIndices{lab6}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
