% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabFour-package.R
\docType{package}
\name{LabFour-package}
\alias{LabFour}
\alias{LabFour-package}
\title{LabFour: Linear Regression}
\description{
Implementation of the linear regression.
}
\author{
\strong{Maintainer}: Aydin Ardalan \email{aydin.ardalan@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
