% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong.R
\name{latlong}
\alias{latlong}
\title{This is a package that interacts with OpenStreetmaps API}
\usage{
latlong(place)
}
\arguments{
\item{place}{(string)}
}
\value{
latitude and longitude
}
\description{
This is a package that interacts with OpenStreetmaps API
}
