\name{LinearRegression}
\alias{LinearRegression}
\alias{LinearRegression}
\docType{package}
\title{
\packageTitle{LinearRegression}
}
\description{
\packageDescription{LinearRegression}
}
\details{
A package to handle linear regression models

}
\author{
\packageAuthor{LinearRegression}

Maintainer: \packageMaintainer{LinearRegression}
}

