% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregRC.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Assign a class representing our model}
\arguments{
\item{formula}{Formula or relation between the two model variables.}

\item{data}{Data to be used to compute the regression parameters.}
}
\description{
This program assigns an RC class which contains the linear regression
parameters and some functions to do a simple linear regression analysis
}

\examples{
data(iris)
linreg_mod <- linreg$new(Petal.Length ~ Species, data = iris)
}
