% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Calculate parameters of a multiple regression model.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{Formula or relation between the two model variables.}

\item{data}{Data to be used to compute the regression parameters.}
}
\value{
The list \code{lr_comps} containing the computation results of
the multiple regression model parameters.
}
\description{
This function computes the linear regression parameters and assigns a class
to the returned object with the computation results.
}
\examples{
data(iris)
lr_comps <- linreg(Petal.Length ~ Species, data = iris)
}
