% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labData.R
\docType{data}
\name{LOINC}
\alias{LOINC}
\alias{refLOINC}
\alias{tagLOINC}
\title{Dataset - LOINC}
\format{These datatable with ICD diagnostic codes and the specific Standardization categories.}
\usage{
refLOINC

tagLOINC
}
\description{
These datasets are used for diagnostic code transformation.
}
\keyword{datasets}
