% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median_sd_fn.R
\name{mean_median_sd}
\alias{mean_median_sd}
\title{Statistics of Average Medicare Payments}
\usage{
mean_median_sd(x)
}
\arguments{
\item{x}{a string, either mean, median or sd (standard deviation)}
}
\value{
Table of of either mean, median or standard deviation
of the average medicare payments by DRG code
}
\description{
This function produces a table of either mean, median or standard deviation
of the average medicare payments by DRG code
}
\examples{
mean_median_sd("sd")


}
