% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labData.R
\docType{data}
\name{MIMIC}
\alias{MIMIC}
\alias{patientMIMIC}
\alias{mapMIMIC}
\alias{labMIMIC}
\title{Dataset - MIMIC}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 46520 rows and 8 columns.}
\usage{
patientMIMIC

mapMIMIC

labMIMIC
}
\description{
Dataset - MIMIC
}
\keyword{datasets}
