% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.goose_race.R
\name{move}
\alias{move}
\alias{move.goose_race}
\title{Method for goose_race class object to move the goose}
\usage{
move(df, x = 1, y = 1)

\method{move}{goose_race}(df, x = 1, y = 1)
}
\arguments{
\item{df}{dataframe to adjust}

\item{x}{integer distance to move along the x axis}

\item{y}{integer distance to move along the y axis}
}
\description{
Method for goose_race class object to move the goose
}
\section{Methods (by class)}{
\itemize{
\item \code{goose_race}: Applied to a goose_race class
}}

\examples{
gr <- goose_race()
move(gr, x = 1, y = 1)

}
