% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{myPrint}
\alias{myPrint}
\title{myPrint (custom print)}
\usage{
myPrint(x, stripoff = FALSE)
}
\arguments{
\item{x}{object}

\item{stripoff}{stripoff column names will be stripped off.}
}
\value{
Nothing
}
\description{
myPrint (custom print)
}
