% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.goose_race.R
\name{plot.goose_race}
\alias{plot.goose_race}
\title{Method for goose_race class object to see a visual of current environment}
\usage{
\method{plot}{goose_race}(x, ...)
}
\arguments{
\item{x}{goose_race object to display}

\item{...}{additional arguments}
}
\description{
Method for goose_race class object to see a visual of current environment
}
