% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labFun.R, R/plotWindowProportion.R
\name{plotWindow}
\alias{plotWindow}
\alias{plotWindowProportion}
\title{Plot Time-Series Window Proportion}
\usage{
plotWindowProportion(
  labData,
  idColName,
  labItemColName,
  dateColName,
  indexDate = last,
  gapDate = c(30, 90, 180, 360),
  topN = 10,
  studyPeriodStartDays,
  studyPeriodEndDays
)
}
\arguments{
\item{labData}{a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.}

\item{idColName}{the column name that records patient ID in labData.}

\item{labItemColName}{the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add \code{+} operator between column names, e.g., \code{A + B}.}

\item{dateColName}{the column name that records test date in labData. It should be in \code{"YYYYMMDD"/"YYYY-MM-DD"} format.}

\item{indexDate}{the specific date that used for cutting time window. It can be first record (\code{"first"}), last record (\code{"last"}), any single date of interest with \code{"YYYYMMDD"/"YYYY-MM-DD"} format, or (\code{indexTable}), with ID and indexDate mapping table.}

\item{gapDate}{desired period (in days) of each window interval. If \code{NULL}, it will be seen as only one single time window.}

\item{studyPeriodStartDays}{the expected start date of your study period, calculated by \code{indexDate} + \code{studyPeriodStartDays}. For example, if \code{studyPeriodStartDays}=0, the start date of your study period will be the \code{indexDate}.}

\item{studyPeriodEndDays}{the expected end date of your study period, calculated by \code{indexDate} + \code{studyPeriodEndDays}. For example, if \code{studyPeriodEndDays}=360, the end date of your study period will be the \code{indexDate} + 360 days.}
}
\value{
A \code{data.table} with statistical summary. \code{By Individual} means the proportion of individuals do not have any test result, {By Window} means the proportion of time windows do not have test results.
}
\description{
\code{getTimeSeriesLab} seperates original data into multiple time windows, and summarize statistical information.
}
\details{
This function is used for seperating lab data into multiple time windows, and it provides overall statistical information: total count, maximun value, minimun value, mean, nearest record to index date of each time window. If \code{indexDate} is first, then it will be the earliest test date among all the lab tests.
}
\examples{
windowProportion <- plotWindowProportion(labData = labSample,
                                        idColName = SUBJECT_ID,
                                        labItemColName = ITEMID,
                                        dateColName = CHARTTIME,
                                        indexDate = first,
                                        gapDate = c(30, 90, 180, 360),
                                        studyPeriodStartDays=0,
                                        studyPeriodEndDays=360
                                        )
print(windowProportion$graph)
head(windowProportion$missingData)

}
