% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_QR.R
\docType{class}
\name{ridgereg_QR-class}
\alias{ridgereg_QR-class}
\alias{ridgereg_QR}
\title{Ridge Regression with QR decomposition}
\description{
Ridge Regression with QR decomposition
}
\section{Fields}{

\describe{
\item{\code{y_est}}{Estimation of y.}

\item{\code{local_data_name}}{data name.}

\item{\code{local_lambda}}{penalty.}

\item{\code{coefs}}{regression coefficients.}
}}


