\name{Rlab3pack-package}
\alias{Rlab3pack-package}
\alias{Rlab3pack}
\docType{package}
\title{
\packageTitle{Rlab3pack}
}
\description{
\packageDescription{Rlab3pack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rlab3pack}
\packageIndices{Rlab3pack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Rlab3pack}

Maintainer: \packageMaintainer{Rlab3pack}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
}

\keyword{ package }

%\examples{
%~~ simple examples of the most important functions ~~
%}
