% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labData.R
\docType{data}
\name{Sample}
\alias{Sample}
\alias{patientSample}
\alias{mapSample}
\alias{labSample}
\title{Dataset - Sample}
\format{
Sample from MIMIC

An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 5 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1744 rows and 6 columns.
}
\usage{
patientSample

mapSample

labSample
}
\description{
Dataset - Sample
}
\keyword{datasets}
