\name{searchapi-package}
\alias{searchapi-package}
\alias{searchapi}
\docType{package}
\title{
\packageTitle{searchapi}
}
\description{
\packageDescription{searchapi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{searchapi}
\packageIndices{searchapi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{searchapi}

Maintainer: \packageMaintainer{searchapi}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }


