% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labFun.R, R/searchCasesByLOINC.R
\name{searchCases}
\alias{searchCases}
\alias{searchCasesByLOINC}
\title{Search cases}
\usage{
searchCasesByLOINC(
  labData,
  idColName = NULL,
  loincColName = NULL,
  dateColName = NULL,
  condition,
  isSummary = FALSE
)
}
\arguments{
\item{labData}{a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.}

\item{idColName}{the column name that records patient ID in labData.}

\item{loincColName}{tthe column name that records LOINC in labData.}

\item{dateColName}{the column name that records test date in labData. It should be in \code{"YYYYMMDD"/"YYYY-MM-DD"} format.}

\item{condition}{desired keyword string.}

\item{isSummary}{summary T or F}
}
\value{
A \code{data.table} merged two tables.
}
\description{
\code{searchCasesByLOINC}
}
\details{
LOINC provides a column to record related names for lab tests. This function can be used to search this kind of description. Only searching LOINC codes are also avaliable, but not recoommended.
}
\examples{
loincSample <- mapLOINC(labData = labSample,
                       labItemColName = ITEMID,
                       mappingTable = mapSample)
searchCasesByLOINC(labData = loincSample,
                  idColName = SUBJECT_ID,
                  loincColName = LOINC,
                  dateColName = CHARTTIME,
                  condition = "Creatinine",
                  isSummary = TRUE)

}
