% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2function.R
\name{stats}
\alias{stats}
\title{Statistics over all of the DRG codes for average Medicare payments}
\usage{
stats(x)
}
\arguments{
\item{x}{a string for type of statistics calculated}
}
\value{
Statistics for average Medicare payments
}
\description{
This function calculates statistics over all of the DRG codes for average Medicare payments.
}
\examples{
stats('mean')

}
