% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistic.R
\name{summary_statistic}
\alias{summary_statistic}
\title{Summary Statistic}
\usage{
summary_statistic(df, FUN)
}
\arguments{
\item{df}{a dataframe}

\item{FUN}{a statistical function; i.e. mean, median, standard deviation}
}
\value{

}
\description{
Summary Statistic
}
\examples{
DRG <- read_csv('DRG_data.csv')

summary_statistic(DRG, mean)

}
