% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team5_BL.R
\name{team5_BL}
\alias{team5_BL}
\alias{team_5}
\title{team5_BL}
\usage{
team5_BL(file,tolerance)
}
\arguments{
\item{file}{A path to the .shp file of a selected country. Default value is Australia.}

\item{tolerance}{A numeric number of thinning parameter. Default value is 0.1.}
}
\value{
A dataframe for plotting the map of the selected country (with longitude, latitude, and grouping indeces of states/provinces). A data frame of the geometry file \code{file}
}
\description{
Function for Lab3 using Lab2 Team5 function
}
\examples{
file <- "inst/gadm36_AU_shp/gadm36_AUS_1.shp"
tolerance <- 0.1
team_5(file = file, tolerance = tolerance)

fpath <- system.file("extdata","gadm36_AUS_1.shp",package="Lab3R")
team5_BL(fpath,0.1)
}
