% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_1.R
\name{team_1}
\alias{team_1}
\title{Function from Lab2 Team1}
\usage{
team_1(file = system.file("extdata", "gadm36_AUS_1.shp", package =
  "Lab3R"), tolerance = 0.1)
}
\arguments{
\item{file}{A path to the .shp file of a selected country. Default value is Australia.}

\item{tolerance}{A numeric number of thinning parameter. Default value is 0.1.}
}
\value{
A dataframe for plotting the map of the selected country (with longitude, latitude, and grouping indeces of states/provinces).
}
\description{
Function from Lab2 Team1
}
\examples{
file = system.file("extdata", "gadm36_AUS_1.shp", package = "Lab3R")
tolerance = 0.1
team_1(file = file, tolerance = tolerance)
}
