% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{team_7}
\alias{team_7}
\title{team_7}
\usage{
team_7(file,tolerance)
}
\arguments{
\item{file}{A path to a shape file}

\item{tolerance}{The value used to thin the polygon}
}
\value{
A data frame of the geometry file \code{file}
}
\description{
Function for Lab3 using Lab2 Team5 function
}
\examples{
fpath <- system.file("extdata","gadm36_AUS_1.shp",package="Lab3R")
team_7(fpath,0.1)
}
