% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{textInputRow}
\alias{textInputRow}
\title{find crime from type}
\usage{
textInputRow(inputId, label, value = "", ...)
}
\arguments{
\item{inputId}{inputId}

\item{label}{label}

\item{value}{value}

\item{...}{other input}

\item{textInputRow}{textInputRow}
}
\description{
find crime from type
}
