% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labFun.R, R/wideTimeSeriesLab.R
\name{wideTS}
\alias{wideTS}
\alias{wideTimeSeriesLab}
\title{Wide Time-Series Data}
\usage{
wideTimeSeriesLab(
  labData,
  idColName,
  labItemColName,
  windowColName,
  valueColName
)
}
\arguments{
\item{labData}{a file or dataframe of laboratory test data with at least 4 columns about patient ID, lab item, test value and test date, respectively.}

\item{idColName}{the column name that records patient ID in labData.}

\item{labItemColName}{the column name that records lab item in labData. If lab code is combined by multiple columns, then just simply add \code{+} operator between column names, e.g., \code{A + B}.}

\item{windowColName}{the column name that records time window sequence in labData.}
}
\value{
A new, wide-formatted \code{data.table}.
}
\description{
\code{wideTimeSeriesLab} transforms time-series data into wide format.
}
\details{
After setting desired time window, data can further transform into wide format, which is more suitable for analysis. If lab item code is composed by multiple columns, it will automaticlly combined as one. And there can only be single value column selected in this step.
}
\examples{

timeSeriesData <- getTimeSeriesLab(labData = labMIMIC,
                                   idColName = SUBJECT_ID,
                                   labItemColName = ITEMID,
                                   dateColName = CHARTTIME,
                                   valueColName = VALUENUM,
                                   indexDate = first,
                                   gapDate = 360,
                                   completeWindows = TRUE)
imputedData <- imputeTimeSeriesLab(labData = timeSeriesData,
                   idColName = ID,
                   labItemColName = ITEMID,
                   windowColName = Window,
                   valueColName = Max & Min,
                   impMethod = mean)
wideTimeSeriesLab(labData = imputedData,
                  idColName = ID,
                  labItemColName = ITEMID,
                  windowColName = Window,
                  valueColName = Max)
}
