% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example graph .}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{vector1, first nodes nodes}
  \item{v2}{vector2, second nodes vector}
  \item{w}{cost from v1 to v2 nodes}
}}
\source{
\url{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}
}
\usage{
wiki_graph
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
