% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Knapsack Solver}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data frame containing two columns: \code{v} (values) and \code{w} (weights).}

\item{W}{A numeric value representing the maximum capacity (weight) of the knapsack.}
}
\value{
A list with the following components:
\item{max_value}{The maximum value that can be obtained.}
\item{selected_items}{The indices of the items selected to obtain the maximum value.}
}
\description{
Solves the 0/1 knapsack problem using a brute force approach, where every
possible subset of items is checked to determine the optimal selection.
}
