% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Calculate the shortest path between nodes in a graph}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Data frame with the edge nodes and the edge distance}

\item{init_node}{Initial node to compute the distances}
}
\value{
The vector \code{dist} containing the minimum distance from the
initial node to the rest of nodes of the graph..
}
\description{
This function computes the shortest path between nodes in a graph using 
the Dijkstra algorithm.
For specifics about the algorithm, refer to the wikipedia page
\url{https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm}
}
\examples{
dijkstra(graph = data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)), init_node = 1)
}
