% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_group11.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm - Find the shortest path of the given graph.}
\usage{
dijkstra(df, init_nod)
}
\arguments{
\item{df}{A data frame (the graph).}

\item{init_nod}{A numbe (the start node).}
}
\value{
The shortest path to every other node from the starting node as a vector.
}
\description{
Dijkstra's algorithm - Find the shortest path of the given graph.
}
\examples{
wiki_graph <- 
data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6), 
           v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5), 
           w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
