% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Implementation of Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame with 3 columns:
\describe{
\item{\code{v1}}{Starting vertex of an edge (numeric).}
\item{\code{v2}}{Ending vertex of an edge (numeric).}
\item{\code{w}}{Weight of the edge (non-negative numeric).}
}}

\item{init_node}{A numeric scalar representing the starting vertex.}
}
\value{
A numeric vector of shortest distances from the starting vertex to each vertex.
}
\description{
This function computes the shortest paths from a starting vertex to all other vertices
in a weighted graph using Dijkstra's algorithm.
}
\examples{
# Example dataset
wiki_graph <- data.frame(
  v1 = c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2 = c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w  = c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)

dijkstra(wiki_graph, 1)
# Expected output: 0 7 9 20 20 11

}
\references{
Wikipedia contributors. \emph{Dijkstra algorithm}. Wikipedia. \url{https://en.wikipedia.org/wiki/Dijkstra_algorithm}
}
