% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Greatest Common Divisor}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
the Greatest common divisor between two numbers
}
\description{
Finds the greatest common divisor of two numbers
}
\examples{
euclidean(123612,13892347912)
euclidean(100, 1000)
}
\references{
https://en.wikipedia.org/wiki/Euclidean algorithm
}
