% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate the gcd of two numbers using the euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Number a}

\item{b}{Number b}
}
\value{
The gcd of \code{a} and \code{b}.
}
\description{
This functions calculates the greatest common denominator (gcd) of two numbers
using the euclidean algorithm.
For specifics about the algorithm, refer to the wikipedia page
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
euclidean(1000, 100)
euclidean(250, 55)
}
