% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Implementation of Greatest Common Divisor using Euclidean Algorithm}
\usage{
euclidean(i, j)
}
\arguments{
\item{i}{an integer}

\item{j}{an integer}
}
\description{
This function calculates the greatest common divisor of two input integers
}
\details{
- If both inputs are zero, the GCD is undefined and an error is raised.
- If one input is zero, the GCD is the absolute value of the other input.
- Negative inputs are allowed; the result is always returned as a positive integer.
}
\examples{
euclidean(120, 25) # Returns 5
euclidean(123612, 13892347912) # Returns 4
euclidean(0, 0) # Returns error
euclidean(0, 12) # Returns 12

}
