% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Implementation of Greatest Common Divisor using Euclidean Algorithm}
\usage{
euclidean(i, j)
}
\arguments{
\item{i}{an integer (numeric scalar)}

\item{j}{an integer (numeric scalar)}
}
\value{
A single numeric value: the greatest common divisor of the two inputs.
}
\description{
This function calculates the greatest common divisor of two input integers
}
\details{
\itemize{
\item If both inputs are zero, then GCD is undefined and an error is raised.
\item If one input is zero, the GCD is the absolute value of the other input.
\item Negative inputs are allowed; the result is always returned as a positive integer.
}
}
\examples{
euclidean(120, 25) # Returns 5
euclidean(123612, 13892347912) # Returns 4
euclidean(0, 12) # Returns 12

}
\references{
Wikipedia Reference \emph{Euclidean algorithm}
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
