% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.r
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{knapsack objects - data frame}

\item{W}{capacity of knapsack - integer}
}
\description{
Greedy heuristic
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(
  w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
)

greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)
}
