% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_access.R
\docType{class}
\name{kolada_api-class}
\alias{kolada_api-class}
\alias{kolada_api}
\title{KOLADA API Reference Class Object}
\arguments{
\item{muni}{data frame of municipality}

\item{kpi}{data frame of KPI}
}
\description{
KOLADA API Reference Class Object
}

\examples{
kolada_mod <- kolada_api$new()
kolada_mod$search_with_title(search_type = "municipality", input_str = "lund")
kolada_mod$search_with_id(search_type = "municipality_groups", input_str = "G124026")
kolada_mod$search_data(input_kpi="N00945",input_municipality="1860",input_year="")
kolada_mod$search_ou("N15033,N15030","V15E144001301,V15E144001101","2009,2008,2007")
}
