% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labData.R
\docType{package}
\name{lab-package}
\alias{lab}
\alias{lab-package}
\title{lab: What the Package Does (Title Case)}
\description{
"dxpr" is a health data analysis tool for ICD diagnosis and procedure codes in electric medical record. It is used for data integration, data wrangling and visualization.
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/ChunJuC/lab}
}

}
\author{
\strong{Maintainer}: Chen-Ju Chen \email{ju.chin.0902@gmail.com}

Authors:
\itemize{
  \item Yi-Ju Tseng \email{yjtseng@mail.cgu.edu.tw}
}

}
\keyword{internal}
