% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3-package.R
\docType{package}
\name{lab3}
\alias{lab3}
\alias{lab3-package}
\title{lab3: A package for calculating dijkstra and euclidean algorithms}
\description{
The lab3 package provides two important functions:
dijkstra, euclidean
}
\section{lab3 functions}{

dijkstra, euclidean
}

