\name{Lab4Package-package}
\alias{Lab4Package-package}
\alias{Lab4Package}
\docType{package}
\title{
\packageTitle{Lab4Package}
}
\description{
\packageDescription{Lab4Package}
}
\author{
Nikolaus Stratil-Sauer and Juan Manuel Pardo Ladino
}

\references{
\url{https://stackoverflow.com/questions/11738844/qpdf-exe-for-compactpdf}

}
\examples{
data(iris)
lr_comps <- linreg(Petal.Length ~ Species, data = iris)
}
