% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabFour-package.R
\docType{package}
\name{LabFour-package}
\alias{LabFour}
\alias{LabFour-package}
\title{LabFour: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
