% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{name <- "Aydin Ardalan , Seyed Mehdi Mirshojaei"
liuid <- "Aydar442, Seymi176"
github_username <- "aydinardalan, Seymi176"
Create a linear regression object}
\arguments{
\item{formula, }{The linear model's formula.}

\item{data, }{The provided dataset.}
}
\value{
empty
}
\description{
name <- "Aydin Ardalan , Seyed Mehdi Mirshojaei"
liuid <- "Aydar442, Seymi176"
github_username <- "aydinardalan, Seymi176"
Create a linear regression object
}
\section{Fields}{

\describe{
\item{\code{formula,}}{Formula for linear model .}

\item{\code{data,}}{The linear model's formula.}

\item{\code{X,}}{matrix. The matrix of independent variables.}

\item{\code{Y,}}{matrix. The matrix of dependent variables.}

\item{\code{Xt,}}{transpose of m_X}

\item{\code{XtX,}}{inverse of matrix multiplication of m_X and X_t}

\item{\code{B,}}{matrix. Regressions coefficients.}

\item{\code{Yfit,}}{The fitted values for y.}

\item{\code{E,}}{The residuals.}

\item{\code{sigsq,}}{The residual variance.}

\item{\code{Estd,}}{#The residual standard deviation.}

\item{\code{N,}}{The number of observations in the model.}

\item{\code{P,}}{The number of parameters in the model.}

\item{\code{var_B,}}{The variance of the regression coefficients.}

\item{\code{tval,}}{The t-values for each coefficient.}

\item{\code{pval,}}{p-values computed according to t-vales and pt function.}

\item{\code{Formula,}}{The exporting formula.}

\item{\code{Data,}}{The exporting data.}

\item{\code{sqrtstresiduals,}}{The square root of standardized residuals.}
}}

\section{Methods}{

\describe{
\item{\code{plot()}}{plots fitted values versus residuals and standardized residuals}
}}

