% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression}
\value{
A class with multiple fields and methods to use on.
}
\description{
Linear Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{defines a formula of how we want to model the relationship between different variables.}

\item{\code{data}}{It is a dataframe containing multiple variables in the columns and different observations of those variables in the rows.}

\item{\code{coeff}}{Regressions coefficients.}

\item{\code{fitted}}{The fitted values.}

\item{\code{residuals}}{The residuals.}

\item{\code{dof}}{The degrees of freedom.}

\item{\code{rvariance}}{The residual variance.}

\item{\code{varcoeff}}{The variance of the regression coefficients.}

\item{\code{tvalues}}{The t-values for each coefficient.}
}}


