% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregRC.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{This is a class representing our model}
\description{
This assigns a RC class
to the returned object with the computation results.
}

