% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.r
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg RC Class Object}
\arguments{
\item{formula}{formula of regression model}

\item{data}{data of regression model}
}
\description{
linreg RC Class Object
}

\examples{
data(iris)
l1 <- linreg$new(formula = Petal.Length~Sepal.Width+Sepal.Length, data = iris)
l1$print()
}
