% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labData.R
\docType{data}
\name{LOINC}
\alias{LOINC}
\alias{refLOINC}
\alias{tagLOINC}
\title{Dataset - LOINC}
\format{
These datatable with ICD diagnostic codes and the specific Standardization categories.

An object of class \code{data.table} (inherits from \code{data.frame}) with 91388 rows and 2 columns.
}
\usage{
refLOINC

tagLOINC
}
\description{
These datasets are used for diagnostic code transformation.
}
\keyword{datasets}
