% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Create a ridge regression object}
\source{
Read more at \url{https://en.wikipedia.org/wiki/Ridge_regression}
}
\arguments{
\item{formula, }{The model's formula.}

\item{data, }{The provided dataset.}
}
\description{
Create a ridge regression object
}
\section{Fields}{

\describe{
\item{\code{formula,}}{Formula for ridge regression .}

\item{\code{data,}}{The provided dataset.}

\item{\code{X,}}{the normalized of the matrix of independent variables.}

\item{\code{my_data_x,}}{The matrix of independent variables.}

\item{\code{y,}}{The matrix of dependent variables.}

\item{\code{yhat,}}{The matrix of dependent variables.}

\item{\code{BhatR,}}{Regressions coefficients.}

\item{\code{lambda,}}{Penalty}

\item{\code{data_name,}}{The exporting data.}
}}


