% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{team_7}
\alias{team_7}
\title{team_7}
\usage{
team_7(file,tolerance)
}
\arguments{
\item{file}{A path to the .shp file of a selected country. Default value is Australia.}

\item{tolerance}{A numeric number of thinning parameter. Default value is 0.1.}
}
\value{
A dataframe for plotting the map of the selected country (with longitude, latitude, and grouping indeces of states/provinces).
}
\description{
Function for Lab3 using Lab2 Team5 function
}
\examples{
fpath <- system.file("extdata","gadm36_AUS_1.shp",package="Lab3R")
team_7(fpath,0.1)
}
