% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{team_7}
\alias{team_7}
\title{team_7}
\usage{
team_7(file,tolerance)
}
\arguments{
\item{file}{A path to a shape file}

\item{tolerance}{The value used to thin the polygon}
}
\value{
A data frame of the geometry file \code{file}
}
\description{
team_7
}
\examples{
tolerance <- 0.1
file <- "./data/gadm36_AUS_shp/gadm36_AUS_1.shp"
team_7(file,tolerance)
}
