\name{wiki_graph}
\alias{wiki_graph}
\docType{data}
\title{wiki_graph data for djikstra algorithm}
\description{wiki_graph data for djikstra algorithm
}
\usage{data("wiki_graph")}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{v1}}{a numeric vector}
    \item{\code{v2}}{a numeric vector}
    \item{\code{w}}{a numeric vector}
  }
}

\references{https://en.wikipedia.org/wiki/Graph
}

