% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph-data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{This data set provides an example graph to easily test the dijkstra() function}
\format{
A data frame with 18 rows and 3 variables:
\describe{
 \item{v1}{starting point of the edge}
 \item{v2}{end point of the edge}
 \item{w}{weight of the edge}
}
}
\source{
The graph was provided in the assignment for this lab.
}
\usage{
data(wiki_graph)
}
\description{
This data set provides an example graph to easily test the dijkstra() function
}
\keyword{datasets}
