% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_group11.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{WIKI GRAPH}
\format{A data frame with three variables:
\describe{
  \item{v1}{start of the edges}
  \item{v2}{end of the edges}
  \item{w}{weight of the edges}
  
}}
\source{
\url{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}
}
\usage{
wiki_graph
}
\description{
A data.frame contains the edges of the graph (from v1 to v2) with the weight of the edge (w).
diamonds.
}
\keyword{datasets}
