% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia Graph Dataset}
\format{
A data frame with 3 columns:
\describe{
\item{v1}{Source node of the edge (numeric).}
\item{v2}{Target node of the edge (numeric).}
\item{w}{Weight of the edge (numeric).}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}
}
\usage{
wiki_graph
}
\description{
A dataset representing a small weighted graph from Wikipedia.
}
\examples{
head(wiki_graph)
dijkstra(wiki_graph, 1)
}
\keyword{datasets}
